pyproject.toml


[project]
name = "s2dm"
dynamic = ["version"]
description = "An friendly approach for modeling semantic data ."
authors = [{name = "Daniel Alvarez-Coello", email="8550265+jdacoello@users.noreply.github.com"}]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
    "click>=8.1.7",
    "rich-click>=1.8.3",
    "rich>=13.9.4",
    "graphql-core>=3.2.6",
    "pyyaml>=6.0.2",
]
requires-python = ">=3.11"

[project.urls]
homepage = "https://github.com/user/s2dm"
documentation = "https://github.com/user/s2dm/blob/main/README.md"

[project.scripts]
s2dm = "s2dm.cli:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/s2dm/__init__.py"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.3",
    "pre-commit>=4.0.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
    "nox>=2024.10.9",
    "bump-my-version>=0.28.1"
]
[tool.mypy]
files = ["src/s2dm"]
show_error_codes = true
strict = true
enable_error_code = ["redundant-expr", "truthy-bool"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PGH", "B", "SIM"]

[tool.bumpversion]
current_version = "0.1.0"

[[tool.bumpversion.files]]
filename = "./src/s2dm/__init__.py"
